back-end/requirements.txt
django==3.2.16
django-environ==0.9.0
django-cors-headers==3.13.0
djangorestframework==3.14.0

argon2-cffi==21.3.0
djoser==2.2.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.5.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2

pillow==9.3.0
django-storages==1.13.1
django-ckeditor==6.3.2
psycopg2==2.9.5

python-slugify==8.0.1

django-cloudinary-storage==0.3.0